+++ hermes/cmd/cmd.go
@@ -45,6 +45,20 @@ func GetCommands() []*plugin.Command {
 							Short: "Start the Hermes relayer",
 						},
 						{
+							Use:   "clear [command]",
+							Short: "Clear the Hermes directories",
+							Commands: []*plugin.Command{
+								{
+									Use:   "binaries",
+									Short: "Clear all binary files from the cache path",
+								},
+								{
+									Use:   "configs",
+									Short: "Clear all configuration files from the config path",
+								},
+							},
+						},
+						{
 							Use:   "keys [command]",
 							Short: "Start the Hermes relayer",
 							Commands: []*plugin.Command{