+++ hermes/pkg/hermes/hermes_test.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 
 	"github.com/ignite/cli/v29/ignite/pkg/errors"
+
 	"github.com/stretchr/testify/require"
 )